Fix library version when linked with sbt via github url
This change fixes a problem found when including the Scrava library into a parent project
via the github url using a project/Build.scala file. (utilising the dependsOn sbt command)

When included this way during the sbt build process the git describe command is run from
the project root directory, not the Scrava library location. Consequently the git command
fails if the parent project is not a git repo. If the parent is a git repo I assume it
will return the version of the parent project instead of the Scrava project.

The change ensures the git command is run from the Scrava project location, not the
parent project.